requirements.txt
streamlit==1.41.1
python-dotenv==0.21.1
langchain==0.2.17
langchain-google-genai==1.0.10
langchain-community==0.2.19
faiss-cpu==1.9.0.post1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.10.5